# openapi.json
{"openapi":"3.1.0","info":{"title":"Agent IA de formation","description":"MVP — formation et évaluation pédagogique (RAG local via Ollama)","version":"0.1.0"},"paths":{"/api/health":{"get":{"tags":["health"],"summary":"Health","operationId":"health_api_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/auth/login":{"post":{"tags":["auth"],"summary":"Login Form","operationId":"login_form_api_auth_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_form_api_auth_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/login/json":{"post":{"tags":["auth"],"summary":"Login Json","operationId":"login_json_api_auth_login_json_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/refresh":{"post":{"tags":["auth"],"summary":"Refresh","operationId":"refresh_api_auth_refresh_post","requestBody":{"content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Body"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/me":{"get":{"tags":["auth"],"summary":"Me","operationId":"me_api_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserOut"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/documents":{"get":{"tags":["documents"],"summary":"List Documents","operationId":"list_documents_api_documents_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/DocumentOut"},"type":"array","title":"Response List Documents Api Documents Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/documents/upload":{"post":{"tags":["documents"],"summary":"Upload Document","operationId":"upload_document_api_documents_upload_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_document_api_documents_upload_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DocumentOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/documents/{document_id}":{"delete":{"tags":["documents"],"summary":"Delete Document","operationId":"delete_document_api_documents__document_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"document_id","in":"path","required":true,"schema":{"type":"string","title":"Document Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/chat/conversations":{"get":{"tags":["chat"],"summary":"List Conversations","operationId":"list_conversations_api_chat_conversations_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ConversationOut"},"type":"array","title":"Response List Conversations Api Chat Conversations Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"delete":{"tags":["chat"],"summary":"Delete All Conversations","operationId":"delete_all_conversations_api_chat_conversations_delete","responses":{"204":{"description":"Successful Response"}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/chat/conversations/{conversation_id}":{"delete":{"tags":["chat"],"summary":"Delete Conversation","operationId":"delete_conversation_api_chat_conversations__conversation_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"conversation_id","in":"path","required":true,"schema":{"type":"string","title":"Conversation Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/chat/conversations/{conversation_id}/messages":{"get":{"tags":["chat"],"summary":"List Messages","operationId":"list_messages_api_chat_conversations__conversation_id__messages_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"conversation_id","in":"path","required":true,"schema":{"type":"string","title":"Conversation Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MessageOut"},"title":"Response List Messages Api Chat Conversations  Conversation Id  Messages Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/chat":{"post":{"tags":["chat"],"summary":"Chat","operationId":"chat_api_chat_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/exercises":{"get":{"tags":["exercises"],"summary":"List Exercises","operationId":"list_exercises_api_exercises_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ExerciseOut"},"type":"array","title":"Response List Exercises Api Exercises Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/exercises/attempts/me":{"get":{"tags":["exercises"],"summary":"My Attempts","operationId":"my_attempts_api_exercises_attempts_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/AttemptOut"},"type":"array","title":"Response My Attempts Api Exercises Attempts Me Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/exercises/{exercise_id}":{"get":{"tags":["exercises"],"summary":"Get Exercise","operationId":"get_exercise_api_exercises__exercise_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"exercise_id","in":"path","required":true,"schema":{"type":"string","title":"Exercise Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExerciseOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/exercises/generate":{"post":{"tags":["exercises"],"summary":"Generate Exercise","operationId":"generate_exercise_api_exercises_generate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateExerciseRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExerciseOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/exercises/{exercise_id}/attempts":{"post":{"tags":["exercises"],"summary":"Submit Attempt","operationId":"submit_attempt_api_exercises__exercise_id__attempts_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"exercise_id","in":"path","required":true,"schema":{"type":"string","title":"Exercise Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubmitAttemptRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AttemptOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/dashboard/learner":{"get":{"tags":["dashboard"],"summary":"Learner Dashboard","operationId":"learner_dashboard_api_dashboard_learner_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LearnerStats"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/dashboard/trainer":{"get":{"tags":["dashboard"],"summary":"Trainer Dashboard","operationId":"trainer_dashboard_api_dashboard_trainer_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TrainerStats"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/dashboard/trainer/export.csv":{"get":{"tags":["dashboard"],"summary":"Export Trainer Csv","operationId":"export_trainer_csv_api_dashboard_trainer_export_csv_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/languages/grammar":{"post":{"tags":["languages"],"summary":"Grammar Check","operationId":"grammar_check_api_languages_grammar_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GrammarRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GrammarResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/languages/comprehension":{"post":{"tags":["languages"],"summary":"Comprehension","operationId":"comprehension_api_languages_comprehension_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ComprehensionRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/languages/pronunciation":{"post":{"tags":["languages"],"summary":"Pronunciation","operationId":"pronunciation_api_languages_pronunciation_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_pronunciation_api_languages_pronunciation_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AttemptOut":{"properties":{"id":{"type":"string","title":"Id"},"exercise_id":{"type":"string","title":"Exercise Id"},"score":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Score"},"max_score":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Max Score"},"feedback":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Feedback"},"weak_topics":{"anyOf":[{"items":{},"type":"array"},{"type":"null"}],"title":"Weak Topics"},"duration_seconds":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Duration Seconds"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","exercise_id","score","max_score","feedback","weak_topics","duration_seconds","created_at"],"title":"AttemptOut"},"Body_login_form_api_auth_login_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"^password$"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","format":"password","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"format":"password","title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_form_api_auth_login_post"},"Body_pronunciation_api_languages_pronunciation_post":{"properties":{"reference_text":{"type":"string","title":"Reference Text"},"audio":{"anyOf":[{"type":"string","contentMediaType":"application/octet-stream"},{"type":"null"}],"title":"Audio"}},"type":"object","required":["reference_text"],"title":"Body_pronunciation_api_languages_pronunciation_post"},"Body_upload_document_api_documents_upload_post":{"properties":{"title":{"type":"string","title":"Title"},"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"}},"type":"object","required":["title","file"],"title":"Body_upload_document_api_documents_upload_post"},"ChatRequest":{"properties":{"message":{"type":"string","minLength":1,"title":"Message"},"conversation_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Conversation Id"},"document_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Document Id"}},"type":"object","required":["message"],"title":"ChatRequest"},"ChatResponse":{"properties":{"conversation_id":{"type":"string","title":"Conversation Id"},"answer":{"type":"string","title":"Answer"},"citations":{"items":{"$ref":"#/components/schemas/Citation"},"type":"array","title":"Citations"}},"type":"object","required":["conversation_id","answer","citations"],"title":"ChatResponse"},"Citation":{"properties":{"document_id":{"type":"string","title":"Document Id"},"document_title":{"type":"string","title":"Document Title"},"chunk_index":{"type":"integer","title":"Chunk Index"},"excerpt":{"type":"string","title":"Excerpt"}},"type":"object","required":["document_id","document_title","chunk_index","excerpt"],"title":"Citation"},"ComprehensionRequest":{"properties":{"document_id":{"type":"string","title":"Document Id"},"question_count":{"type":"integer","maximum":10.0,"minimum":1.0,"title":"Question Count","default":3}},"type":"object","required":["document_id"],"title":"ComprehensionRequest"},"ConversationOut":{"properties":{"id":{"type":"string","title":"Id"},"title":{"type":"string","title":"Title"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","title","created_at"],"title":"ConversationOut"},"DocumentOut":{"properties":{"id":{"type":"string","title":"Id"},"title":{"type":"string","title":"Title"},"filename":{"type":"string","title":"Filename"},"status":{"type":"string","title":"Status"},"error_message":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error Message"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","title","filename","status","created_at"],"title":"DocumentOut"},"ExerciseOut":{"properties":{"id":{"type":"string","title":"Id"},"title":{"type":"string","title":"Title"},"exercise_type":{"type":"string","title":"Exercise Type"},"topic":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Topic"},"payload":{"additionalProperties":true,"type":"object","title":"Payload"},"time_limit_seconds":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Time Limit Seconds"},"document_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Document Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","title","exercise_type","topic","payload","time_limit_seconds","document_id","created_at"],"title":"ExerciseOut"},"GenerateExerciseRequest":{"properties":{"document_id":{"type":"string","title":"Document Id"},"exercise_type":{"type":"string","title":"Exercise Type"},"title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title"},"topic":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Topic"},"question_count":{"type":"integer","maximum":20.0,"minimum":1.0,"title":"Question Count","default":5},"time_limit_seconds":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Time Limit Seconds"}},"type":"object","required":["document_id","exercise_type"],"title":"GenerateExerciseRequest"},"GrammarRequest":{"properties":{"text":{"type":"string","minLength":1,"title":"Text"},"language":{"type":"string","title":"Language","default":"fr"}},"type":"object","required":["text"],"title":"GrammarRequest"},"GrammarResponse":{"properties":{"corrected_text":{"type":"string","title":"Corrected Text"},"explanations":{"items":{"type":"string"},"type":"array","title":"Explanations"}},"type":"object","required":["corrected_text","explanations"],"title":"GrammarResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LearnerStats":{"properties":{"attempts_count":{"type":"integer","title":"Attempts Count"},"average_score":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Average Score"},"documents_available":{"type":"integer","title":"Documents Available"},"weak_topics":{"items":{"type":"string"},"type":"array","title":"Weak Topics"},"recent_attempts":{"items":{"$ref":"#/components/schemas/AttemptOut"},"type":"array","title":"Recent Attempts"}},"type":"object","required":["attempts_count","average_score","documents_available","weak_topics","recent_attempts"],"title":"LearnerStats"},"LoginRequest":{"properties":{"email":{"type":"string","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"LoginRequest"},"MessageOut":{"properties":{"id":{"type":"string","title":"Id"},"role":{"type":"string","title":"Role"},"content":{"type":"string","title":"Content"},"citations":{"anyOf":[{"items":{},"type":"array"},{"type":"null"}],"title":"Citations"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","role","content","created_at"],"title":"MessageOut"},"SubmitAttemptRequest":{"properties":{"answers":{"additionalProperties":true,"type":"object","title":"Answers"},"duration_seconds":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Duration Seconds"}},"type":"object","required":["answers"],"title":"SubmitAttemptRequest"},"TokenResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"refresh_token":{"type":"string","title":"Refresh Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"}},"type":"object","required":["access_token","refresh_token"],"title":"TokenResponse"},"TrainerStats":{"properties":{"learners_count":{"type":"integer","title":"Learners Count"},"documents_count":{"type":"integer","title":"Documents Count"},"indexed_documents":{"type":"integer","title":"Indexed Documents"},"attempts_count":{"type":"integer","title":"Attempts Count"},"average_score":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Average Score"},"recurrent_weak_topics":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Recurrent Weak Topics"},"score_by_exercise_type":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Score By Exercise Type"}},"type":"object","required":["learners_count","documents_count","indexed_documents","attempts_count","average_score","recurrent_weak_topics","score_by_exercise_type"],"title":"TrainerStats"},"UserOut":{"properties":{"id":{"type":"string","title":"Id"},"email":{"type":"string","title":"Email"},"full_name":{"type":"string","title":"Full Name"},"role":{"type":"string","title":"Role"},"organization_id":{"type":"string","title":"Organization Id"}},"type":"object","required":["id","email","full_name","role","organization_id"],"title":"UserOut"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/api/auth/login"}}}}}}